requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-cleanup==8.0.0
django-crispy-forms==2.1
django-filter==23.5
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.6.0
